Move listen_for_shutdown_signals to linera-base
