+++ linera-base/src/lib.rs
@@ -12,6 +12,8 @@ use std::fmt;
 
 #[doc(hidden)]
 pub use async_trait::async_trait;
+#[cfg(not(target_arch = "wasm32"))]
+use {::tracing::debug, tokio::signal::unix, tokio_util::sync::CancellationToken};
 pub mod abi;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod command;
@@ -146,4 +148,24 @@ pub fn hex_vec_debug(list: &Vec<Vec<u8>>, f: &mut fmt::Formatter) -> fmt::Result
         hex_debug(bytes, f)?;
     }
     write!(f, "]")
+}
+
+/// Listens for shutdown signals, and notifies the [`CancellationToken`] if one is
+/// received.
+#[cfg(not(target_arch = "wasm32"))]
+pub async fn listen_for_shutdown_signals(shutdown_sender: CancellationToken) {
+    let _shutdown_guard = shutdown_sender.drop_guard();
+
+    let mut sigint =
+        unix::signal(unix::SignalKind::interrupt()).expect("Failed to set up SIGINT handler");
+    let mut sigterm =
+        unix::signal(unix::SignalKind::terminate()).expect("Failed to set up SIGTERM handler");
+    let mut sighup =
+        unix::signal(unix::SignalKind::hangup()).expect("Failed to set up SIGHUP handler");
+
+    tokio::select! {
+        _ = sigint.recv() => debug!("Received SIGINT"),
+        _ = sigterm.recv() => debug!("Received SIGTERM"),
+        _ = sighup.recv() => debug!("Received SIGHUP"),
+    }
 }
