+++ linera-service/src/util.rs
@@ -16,8 +16,6 @@ use http::Uri;
 use linera_base::command::parse_version_message;
 use linera_base::data_types::TimeDelta;
 pub use linera_client::util::*;
-use tokio::signal::unix;
-use tokio_util::sync::CancellationToken;
 use tracing::debug;
 
 /// Extension trait for [`tokio::process::Child`].
@@ -36,25 +34,6 @@ impl ChildExt for tokio::process::Child {
         }
         debug!("Child process {:?} is running as expected.", self);
         Ok(())
-    }
-}
-
-/// Listens for shutdown signals, and notifies the [`CancellationToken`] if one is
-/// received.
-pub async fn listen_for_shutdown_signals(shutdown_sender: CancellationToken) {
-    let _shutdown_guard = shutdown_sender.drop_guard();
-
-    let mut sigint =
-        unix::signal(unix::SignalKind::interrupt()).expect("Failed to set up SIGINT handler");
-    let mut sigterm =
-        unix::signal(unix::SignalKind::terminate()).expect("Failed to set up SIGTERM handler");
-    let mut sighup =
-        unix::signal(unix::SignalKind::hangup()).expect("Failed to set up SIGHUP handler");
-
-    tokio::select! {
-        _ = sigint.recv() => debug!("Received SIGINT"),
-        _ = sigterm.recv() => debug!("Received SIGTERM"),
-        _ = sighup.recv() => debug!("Received SIGHUP"),
     }
 }
